The ability to edit services is functional
Can be merged into development branch, there are a few changes that can
be made in the front end and backend, Make it impossible to add services
that the employee already has. Have a way to pass employeeId into each
component so that the component does not have to have its own individual
dropdown box.